[CORE][HASHMAP] Add collision handling and set function
The hashing function calculates the index of an object by its
identifier (a string). If there are already objects at the same index,
the function has to take care of the *collision* and navigate through the
*linked list* inside the cell and append the object at the end.

The set function is useful for changing objects, in functions such as
(set! <identifier> <new value>).