[CORE][HASHMAP] Add delete() and some fixes.
I'm facing with a lot of bugs with this implementation of hashmap,
probably I'll add helper functions (such as car() or cdr()) for making the
code more readable. At the moment it's an hell of '->' and dots.

I've added the delete() function for removing objects from hashmaps;
I don't know if it will be useful in future.