[CORE] Implement environment
Environments are used as set of variables and procedures, so basically
everything declared in bulb is contained inside of an environment.

An environment contains an hashmap (which contains local variables and
and a pointer to the 'parent' environment so the search for a variable starts at
the local hashmap and continues through the various levels of environments, same
for procedures.