[CORE] Start working on the Garbage Collector
The Garbage Collector offers the abstraction of an *infinite* memory.
The code can keep allocating stuff without having to deal with the memory
management needed for cleaning when there are some garbage (objects
which are no more referenced) in it.

I've decided to go with a simple mark-and-sweep garbage collector. Basically
it "marks" all the reachable objects inside the global environment and then
"sweeps", removing all the objects which are not reachable.

For more information:
http://journal.stuffwithstuff.com/2013/12/08/babys-first-garbage-collector/