[ALL] Implement the new type system
The new type system reduces a lot the memory consumption by
using an *unsigned char* for holding the type tag, but staying flexible
for adding new types.

The GC added the problem of a custom *free* function for deallocating the
data in an object because it can contain nested data and the function free()
offered by the C's standard library was not designed for deallocating
recursively.

So I came up with the idea of a "type table" containing all the types, with
all the functions necessary (at the moment 'free' and 'toString'). How can
the interpreter lookup in this table? The type tag contains the index.