[MAKEFILE] Simplify and fix for GCC 10
Added -fcommon to list of CFLAGS to fix multiple definition errors.

Changed optimisation to `-Os` and added stripping `-s` to demonstrate
how lightweight the interpreter is.

Install, Uninstall, and Dist options were removed from Makefile as I
don't think there is a good usecase for these functions in this project.

Signed-off-by: Mahyar Koshkouei <mk@deltabeard.com>